Embed versioning information in the library
C (and C++) libraries embed version information, and when shared
objects are distributed, the versioning information is used to ensure
that the library is appropriately compatible with the code it was
built against.

This change makes the library have a version that matches the major
version.  That is, it commits the project to maintaining backward ABI
compatibility unless the major version number changes.

This is concretely useful for any system that distributes library
updates independently from the software (e.g. to fix a security bug in
the underlying library).  For example, this includes most GNU/Linux
distributions.